Created app component
