+++ app/app.component.js
@@ -9,17 +9,18 @@ var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
 var core_1 = require('@angular/core');
-var HomepageComponent = (function () {
-    function HomepageComponent() {
+var AppComponent = (function () {
+    function AppComponent() {
     }
-    HomepageComponent = __decorate([
+    AppComponent = __decorate([
         core_1.Component({
-            selector: 'main-app',
-            template: '<h1>Freelance Bootcamp Homepage</h1>'
+            moduleId: module.id,
+            selector: 'app',
+            template: '<h1>Navigation</h1>'
         }), 
         __metadata('design:paramtypes', [])
-    ], HomepageComponent);
-    return HomepageComponent;
+    ], AppComponent);
+    return AppComponent;
 }());
-exports.HomepageComponent = HomepageComponent;
-//# sourceMappingURL=homepage.component.js.map
+exports.AppComponent = AppComponent;
+//# sourceMappingURL=app.component.js.map
