+++ app/app.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+	moduleId: module.id,
+	selector: 'app',
+	template: '<h1>Navigation</h1>'
+})
+export class AppComponent {
+	title: 'Freelance Bootcamp Dashboard';
+}
