+++ app/app.module.ts
@@ -1,17 +1,19 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
-import { HomepageComponent } from './homepage.component';
+import { AppComponent } from './app.component';
+import { HomepageComponent } from './homepage/homepage.component';
 
 @NgModule({
 	imports: [
 		BrowserModule
 	],
 	declarations: [
+		AppComponent,
 		HomepageComponent
 	],
 	bootstrap: [
-		HomepageComponent
+		AppComponent
 	]
 })
 export class AppModule {}
