+++ app/homepage/homepage.component.js
@@ -9,28 +9,17 @@ var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
 var core_1 = require('@angular/core');
-var platform_browser_1 = require('@angular/platform-browser');
-var app_component_1 = require('./app.component');
-var homepage_component_1 = require('./homepage/homepage.component');
-var AppModule = (function () {
-    function AppModule() {
+var HomepageComponent = (function () {
+    function HomepageComponent() {
     }
-    AppModule = __decorate([
-        core_1.NgModule({
-            imports: [
-                platform_browser_1.BrowserModule
-            ],
-            declarations: [
-                app_component_1.AppComponent,
-                homepage_component_1.HomepageComponent
-            ],
-            bootstrap: [
-                app_component_1.AppComponent
-            ]
+    HomepageComponent = __decorate([
+        core_1.Component({
+            selector: 'homepage',
+            template: '<h1>Home</h1>'
         }), 
         __metadata('design:paramtypes', [])
-    ], AppModule);
-    return AppModule;
+    ], HomepageComponent);
+    return HomepageComponent;
 }());
-exports.AppModule = AppModule;
-//# sourceMappingURL=app.module.js.map
+exports.HomepageComponent = HomepageComponent;
+//# sourceMappingURL=homepage.component.js.map
