+++ app/homepage/homepage.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-	selector: 'main-app',
-	template: '<h1>Freelance Bootcamp Homepage</h1>'
+	selector: 'homepage',
+	template: '<h1>Home</h1>'
 })
 export class HomepageComponent {}
