+++ index.html
@@ -15,6 +15,6 @@
   </head>
 
   <body>
-    <main-app>App Loading...</main-app>
+    <app>App Loading...</app>
   </body>
 </html>
